requirements.txt
# Requirements for video-project
# For getting native language subtitles from videos
git+https://github.com/openai/whisper.git

# For translating subtitles
deep-translator
openai
srt
# google-cloud-translate
argostranslate

# For generating audio from text
git+https://github.com/suno-ai/bark.git
pydub
audioop-lts; python_version>='3.13'

# Environment management
python-dotenv